backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
dj-database-url==1.2.0
mysqlclient==2.1.1
whitenoise==6.3.0
gunicorn==20.1.0
PyJWT==2.6.0
requests==2.31.0
xhtml2pdf==1.4.0
pandas==2.0.0
python-dotenv==1.0.0
pillow==9.5.0
pathlib==1.0.1